packages/judgment/pyproject.toml
[build-system]
requires = ["setuptools>=69.0", "wheel"]
build-backend = "setuptools.backends._legacy:_Backend"

[project]
name = "only-reason-judgment"
version = "0.1.0"
description = "Only Reason 0 Agent — Judgment Service"
requires-python = ">=3.11"
dependencies = [
    "fastapi>=0.115.0",
    "uvicorn[standard]>=0.34.0",
    "asyncpg>=0.30.0",
    "pydantic>=2.10.0",
    "sentence-transformers>=3.3.0",
    "openai-whisper>=20240930",
    "numpy>=1.26.0",
    "python-dotenv>=1.0.0",
    "httpx>=0.28.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0.0",
    "pytest-asyncio>=0.24.0",
    "ruff>=0.8.0",
]

[tool.ruff]
target-version = "py311"
line-length = 100
